requirements.txt
# Streamlit 本体
streamlit
pyyaml

# LangChain 関連
langchain
langchain-openai
langchain-community
langchain-core
langchain_huggingface
pydantic-settings>=2.0

# HuggingFace 関連
transformers
sentence-transformers
torch==2.7.1
accelerate
safetensors

# ベクトルDB
faiss-cpu

# LLM 評価や補助
tiktoken
pydantic

# .env 読み込み
python-dotenv

# 文書処理
python-docx

# その他
tqdm
